Split off Noise tests from test_random.py into test_noise.py (#809c)
