+++ tests/test_random.py
@@ -71,24 +71,6 @@ pMean = 7
 # the right answer for the first three Poisson deviates produced from testseed
 pResult = (4, 5, 6)
 
-# gain, read noise to use for CCD noise tests
-cGain = 3.
-cReadNoise = 5.
-
-# types to use in CCD tests
-types = (np.int16, np.int32, np.float32, np.float64)
-typestrings = ("S", "I", "F", "D")
-
-# constant sky background level to use for CCD noise test image
-sky = 50
-
-# Tabulated results
-cResultS = np.array([[44, 47], [50, 49]], dtype=np.int16)
-cResultI = np.array([[44, 47], [50, 49]], dtype=np.int32)
-cResultF = np.array([[44.45332718, 47.79725266], [50.67744064, 49.58272934]], dtype=np.float32)
-cResultD = np.array([[44.453328440057618, 47.797254142519577],
-                     [50.677442088335162, 49.582730949808081]],dtype=np.float64)
-
 # a & b to use for Weibull tests
 wA = 4.
 wB = 9.
@@ -251,36 +233,10 @@ def test_uniform():
             test_array, 2.*np.array(uResult), precisionF,
             err_msg='Wrong uniform random number sequence from generate.')
 
-    # Test filling an image
-    u.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(u))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(uResult), precision,
-            err_msg='Wrong uniform random number sequence generated when applied to image.')
-
-    # Test filling a single-precision image
-    u.seed(testseed)
-    testimage = galsim.ImageF(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(u))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(uResult), precisionF,
-            err_msg='Wrong uniform random number sequence generated when applied to ImageF.')
-
-    # Test filling an image with Fortran ordering
-    u.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((5, 5)).T)
-    testimage.addNoise(galsim.DeviateNoise(u))
-    np.testing.assert_array_almost_equal(
-            [testimage(1,1),testimage(2,1),testimage(3,1)], np.array(uResult), precision,
-            err_msg="Wrong uniform randoms generated for Fortran-ordered Image")
-
     # Check picklability
     do_pickle(u, lambda x: x.serialize())
     do_pickle(u, lambda x: (x(), x(), x(), x()))
-    do_pickle(galsim.DeviateNoise(u), drawNoise)
     do_pickle(u)
-    do_pickle(galsim.DeviateNoise(u))
 
 
 @timer
@@ -413,167 +369,10 @@ def test_gaussian():
             test_array, np.array(gResult)-gMean, precisionF,
             err_msg='Wrong Gaussian random number sequence from generate_from_variance.')
 
-    # Test filling an image
-    g.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(g))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(gResult), precision,
-            err_msg='Wrong Gaussian random number sequence generated when applied to image.')
-
-    # Test filling a single-precision image
-    g.seed(testseed)
-    testimage = galsim.ImageF(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(g))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(gResult), precisionF,
-            err_msg='Wrong Gaussian random number sequence generated when applied to ImageF.')
-
-    # GaussianNoise is equivalent, but no mean allowed.
-    rng.seed(testseed)
-    gn = galsim.GaussianNoise(rng, sigma=gSigma)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(gn)
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(gResult)-gMean, precision,
-            err_msg="GaussianNoise applied to Images does not reproduce expected sequence")
-
-    # Test filling an image with Fortran ordering
-    rng.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((5, 5)).T)
-    testimage.addNoise(gn)
-    np.testing.assert_array_almost_equal(
-            [testimage(1,1),testimage(2,1),testimage(3,1)], np.array(gResult)-gMean, precision,
-            err_msg="Wrong Gaussian noise generated for Fortran-ordered Image")
-
-    # Check GaussianNoise variance:
-    np.testing.assert_almost_equal(
-            gn.getVariance(), gSigma**2, precision,
-            err_msg="GaussianNoise getVariance returns wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getSigma(), gSigma, precision,
-            err_msg="GaussianNoise getSigma returns wrong value")
-
-    # Check that the noise model really does produce this variance.
-    big_im = galsim.Image(2048,2048,dtype=float)
-    big_im.addNoise(gn)
-    var = np.var(big_im.array)
-    print('variance = ',var)
-    print('getVar = ',gn.getVariance())
-    np.testing.assert_almost_equal(
-            var, gn.getVariance(), 1,
-            err_msg='Realized variance for GaussianNoise did not match getVariance()')
-
-    # Check that GaussianNoise adds to the image, not overwrites the image.
-    gal = galsim.Exponential(half_light_radius=2.3, flux=1.e4)
-    gal.drawImage(image=big_im)
-    big_im.addNoise(gn)
-    gal.withFlux(-1.e4).drawImage(image=big_im, add_to_image=True)
-    var = np.var(big_im.array)
-    np.testing.assert_almost_equal(
-            var, gn.getVariance(), 1,
-            err_msg='GaussianNoise wrong when already an object drawn on the image')
-
-    # Check that DeviateNoise adds to the image, not overwrites the image.
-    gal.drawImage(image=big_im)
-    big_im.addNoise(galsim.DeviateNoise(g))
-    gal.withFlux(-1.e4).drawImage(image=big_im, add_to_image=True)
-    var = np.var(big_im.array)
-    np.testing.assert_almost_equal(
-            var, gn.getVariance(), 1,
-            err_msg='DeviateNoise wrong when already an object drawn on the image')
-
-    # Check that VariableGaussianNoise adds to the image, not overwrites the image.
-    gal.drawImage(image=big_im)
-    var_im = big_im.copy()
-    var_im.fill(gn.getVariance())
-    print('rng = ',rng)
-    print('var_im = ',var_im)
-    big_im.addNoise(galsim.VariableGaussianNoise(rng, var_im));
-    gal.withFlux(-1.e4).drawImage(image=big_im, add_to_image=True)
-    var = np.var(big_im.array)
-    np.testing.assert_almost_equal(
-            var, gn.getVariance(), 1,
-            err_msg='VariabeGaussianNoise wrong when already an object drawn on the image')
-
-    # Check withVariance
-    gn = gn.withVariance(9.)
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 9, precision,
-            err_msg="GaussianNoise withVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getSigma(), 3., precision,
-            err_msg="GaussianNoise withVariance results in wrong sigma")
-
-    # Check withScaledVariance
-    gn = gn.withScaledVariance(4.)
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 36., precision,
-            err_msg="GaussianNoise withScaledVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getSigma(), 6., precision,
-            err_msg="GaussianNoise withScaledVariance results in wrong sigma")
-
-    # Check arithmetic
-    gn = gn.withVariance(0.5)
-    gn2 = gn * 3
-    np.testing.assert_almost_equal(
-            gn2.getVariance(), 1.5, precision,
-            err_msg="GaussianNoise gn*3 results in wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 0.5, precision,
-            err_msg="GaussianNoise gn*3 results in wrong variance for original gn")
-    gn2 = 5 * gn
-    np.testing.assert_almost_equal(
-            gn2.getVariance(), 2.5, precision,
-            err_msg="GaussianNoise 5*gn results in wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 0.5, precision,
-            err_msg="GaussianNoise 5*gn results in wrong variance for original gn")
-    gn2 = gn/2
-    np.testing.assert_almost_equal(
-            gn2.getVariance(), 0.25, precision,
-            err_msg="GaussianNoise gn/2 results in wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 0.5, precision,
-            err_msg="GaussianNoise 5*gn results in wrong variance for original gn")
-    gn *= 3
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 1.5, precision,
-            err_msg="GaussianNoise gn*=3 results in wrong variance")
-    gn /= 2
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 0.75, precision,
-            err_msg="GaussianNoise gn/=2 results in wrong variance")
-
-    # Check starting with GaussianNoise()
-    gn = galsim.GaussianNoise()
-    gn = gn.withVariance(9.)
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 9, precision,
-            err_msg="GaussianNoise().withVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getSigma(), 3., precision,
-            err_msg="GaussianNoise().withVariance results in wrong sigma")
-
-    gn = galsim.GaussianNoise()
-    gn = gn.withScaledVariance(4.)
-    np.testing.assert_almost_equal(
-            gn.getVariance(), 4., precision,
-            err_msg="GaussianNoise().withScaledVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            gn.getSigma(), 2., precision,
-            err_msg="GaussianNoise().withScaledVariance results in wrong sigma")
-
     # Check picklability
     do_pickle(g, lambda x: (x.serialize(), x.getMean(), x.getSigma()))
     do_pickle(g, lambda x: (x(), x(), x(), x()))
-    do_pickle(gn, lambda x: (x.rng.serialize(), x.sigma))
-    do_pickle(gn, drawNoise)
-    do_pickle(galsim.DeviateNoise(g), drawNoise)
     do_pickle(g)
-    do_pickle(gn)
-    do_pickle(galsim.DeviateNoise(g))
 
 
 @timer
@@ -689,20 +488,10 @@ def test_binomial():
             test_array, np.array(bResult), precisionI,
             err_msg='Wrong binomial random number sequence from generate.')
 
-    # Test filling an image
-    b.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(b))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(bResult), precision,
-            err_msg='Wrong binomial random number sequence generated when applied to image.')
-
     # Check picklability
     do_pickle(b, lambda x: (x.serialize(), x.getN(), x.getP()))
     do_pickle(b, lambda x: (x(), x(), x(), x()))
-    do_pickle(galsim.DeviateNoise(b), drawNoise)
     do_pickle(b)
-    do_pickle(galsim.DeviateNoise(b))
 
 
 @timer
@@ -834,144 +623,10 @@ def test_poisson():
             test_array, np.array(pResult), precisionI,
             err_msg='Wrong poisson random number sequence from generate_from_expectation.')
 
-    # Test filling an image
-    p.seed(testseed)
-    testimage = galsim.ImageI(np.zeros((3, 1), dtype=np.int32))
-    testimage.addNoise(galsim.DeviateNoise(p))
-    np.testing.assert_array_equal(
-            testimage.array.flatten(), np.array(pResult),
-            err_msg='Wrong poisson random number sequence generated when applied to image.')
-
-    # The PoissonNoise version also subtracts off the mean value
-    rng.seed(testseed)
-    pn = galsim.PoissonNoise(rng, sky_level=pMean)
-    testimage.fill(0)
-    testimage.addNoise(pn)
-    np.testing.assert_array_equal(
-            testimage.array.flatten(), np.array(pResult)-pMean,
-            err_msg='Wrong poisson random number sequence generated using PoissonNoise')
-
-    # Test filling a single-precision image
-    rng.seed(testseed)
-    testimage = galsim.ImageF(np.zeros((3, 1)))
-    testimage.addNoise(pn)
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(pResult)-pMean, precisionF,
-            err_msg='Wrong Poisson random number sequence generated when applied to ImageF.')
-
-    # Test filling an image with Fortran ordering
-    rng.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((5, 5)).T)
-    testimage.addNoise(pn)
-    np.testing.assert_array_almost_equal(
-            [testimage(1,1),testimage(2,1),testimage(3,1)], np.array(pResult)-pMean,
-            err_msg="Wrong Poisson noise generated for Fortran-ordered Image")
-
-    # Check PoissonNoise variance:
-    np.testing.assert_almost_equal(
-            pn.getVariance(), pMean, precision,
-            err_msg="PoissonNoise getVariance returns wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getSkyLevel(), pMean, precision,
-            err_msg="PoissonNoise getSkyLevel returns wrong value")
-
-    # Check that the noise model really does produce this variance.
-    big_im = galsim.Image(2048,2048,dtype=float)
-    big_im.addNoise(pn)
-    var = np.var(big_im.array)
-    print('variance = ',var)
-    print('getVar = ',pn.getVariance())
-    np.testing.assert_almost_equal(
-            var, pn.getVariance(), 1,
-            err_msg='Realized variance for PoissonNoise did not match getVariance()')
-
-    # Check that PoissonNoise adds to the image, not overwrites the image.
-    gal = galsim.Exponential(half_light_radius=2.3, flux=0.3)
-    # Note: in this case, flux/size^2 needs to be << sky_level or it will mess up the statistics.
-    gal.drawImage(image=big_im)
-    big_im.addNoise(pn)
-    gal.withFlux(-0.3).drawImage(image=big_im, add_to_image=True)
-    var = np.var(big_im.array)
-    np.testing.assert_almost_equal(
-            var, pn.getVariance(), 1,
-            err_msg='PoissonNoise wrong when already an object drawn on the image')
-
-    # Check withVariance
-    pn = pn.withVariance(9.)
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 9., precision,
-            err_msg="PoissonNoise withVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getSkyLevel(), 9., precision,
-            err_msg="PoissonNoise withVariance results in wrong skyLevel")
-
-    # Check withScaledVariance
-    pn = pn.withScaledVariance(4.)
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 36, precision,
-            err_msg="PoissonNoise withScaledVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getSkyLevel(), 36., precision,
-            err_msg="PoissonNoise withScaledVariance results in wrong skyLevel")
-
-    # Check arithmetic
-    pn = pn.withVariance(0.5)
-    pn2 = pn * 3
-    np.testing.assert_almost_equal(
-            pn2.getVariance(), 1.5, precision,
-            err_msg="PoissonNoise pn*3 results in wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 0.5, precision,
-            err_msg="PoissonNoise pn*3 results in wrong variance for original pn")
-    pn2 = 5 * pn
-    np.testing.assert_almost_equal(
-            pn2.getVariance(), 2.5, precision,
-            err_msg="PoissonNoise 5*pn results in wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 0.5, precision,
-            err_msg="PoissonNoise 5*pn results in wrong variance for original pn")
-    pn2 = pn/2
-    np.testing.assert_almost_equal(
-            pn2.getVariance(), 0.25, precision,
-            err_msg="PoissonNoise pn/2 results in wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 0.5, precision,
-            err_msg="PoissonNoise 5*pn results in wrong variance for original pn")
-    pn *= 3
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 1.5, precision,
-            err_msg="PoissonNoise pn*=3 results in wrong variance")
-    pn /= 2
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 0.75, precision,
-            err_msg="PoissonNoise pn/=2 results in wrong variance")
-
-    # Check starting with PoissonNoise()
-    pn = galsim.PoissonNoise()
-    pn = pn.withVariance(9.)
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 9., precision,
-            err_msg="PoissonNoise().withVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getSkyLevel(), 9., precision,
-            err_msg="PoissonNoise().withVariance results in wrong skyLevel")
-    pn = pn.withScaledVariance(4.)
-    np.testing.assert_almost_equal(
-            pn.getVariance(), 36, precision,
-            err_msg="PoissonNoise().withScaledVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            pn.getSkyLevel(), 36., precision,
-            err_msg="PoissonNoise().withScaledVariance results in wrong skyLevel")
-
     # Check picklability
     do_pickle(p, lambda x: (x.serialize(), x.getMean()))
     do_pickle(p, lambda x: (x(), x(), x(), x()))
-    do_pickle(pn, lambda x: (x.rng.serialize(), x.sky_level))
-    do_pickle(pn, drawNoise)
-    do_pickle(galsim.DeviateNoise(p), drawNoise)
     do_pickle(p)
-    do_pickle(pn)
-    do_pickle(galsim.DeviateNoise(p))
 
 
 @timer
@@ -1095,20 +750,10 @@ def test_weibull():
             test_array, np.array(wResult), precisionF,
             err_msg='Wrong weibull random number sequence from generate.')
 
-    # Test filling an image
-    w.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(w))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(wResult), precision,
-            err_msg='Wrong weibull random number sequence generated when applied to image.')
-
     # Check picklability
     do_pickle(w, lambda x: (x.serialize(), x.getA(), x.getB()))
     do_pickle(w, lambda x: (x(), x(), x(), x()))
-    do_pickle(galsim.DeviateNoise(w), drawNoise)
     do_pickle(w)
-    do_pickle(galsim.DeviateNoise(w))
 
 
 @timer
@@ -1222,20 +867,10 @@ def test_gamma():
             test_array, np.array(gammaResult), precisionF,
             err_msg='Wrong gamma random number sequence from generate.')
 
-    # Test filling an image
-    g.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(g))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(gammaResult), precision,
-            err_msg='Wrong gamma random number sequence generated when applied to image.')
-
     # Check picklability
     do_pickle(g, lambda x: (x.serialize(), x.getK(), x.getTheta()))
     do_pickle(g, lambda x: (x(), x(), x(), x()))
-    do_pickle(galsim.DeviateNoise(g), drawNoise)
     do_pickle(g)
-    do_pickle(galsim.DeviateNoise(g))
 
 
 @timer
@@ -1349,20 +984,10 @@ def test_chi2():
             test_array, np.array(chi2Result), precisionF,
             err_msg='Wrong Chi^2 random number sequence from generate.')
 
-    # Test filling an image
-    c.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(c))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(chi2Result), precision,
-            err_msg='Wrong Chi^2 random number sequence generated when applied to image.')
-
     # Check picklability
     do_pickle(c, lambda x: (x.serialize(), x.getN()))
     do_pickle(c, lambda x: (x(), x(), x(), x()))
-    do_pickle(galsim.DeviateNoise(c), drawNoise)
     do_pickle(c)
-    do_pickle(galsim.DeviateNoise(c))
 
 
 @timer
@@ -1527,21 +1152,9 @@ def test_distfunction():
             test_array, 2*np.array(dFunctionResult), precisionF,
             err_msg='Wrong DistDeviate random number sequence from add_generate.')
 
-    # Test filling an image
-    d.seed(testseed)
-    print('d = ',d)
-    print('d._rng = ',d._rng)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(d))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(dFunctionResult), precision,
-            err_msg='Wrong DistDeviate random number sequence generated when applied to image.')
-
     # Check picklability
     do_pickle(d, lambda x: (x(), x(), x(), x()))
-    do_pickle(galsim.DeviateNoise(d), drawNoise)
     do_pickle(d)
-    do_pickle(galsim.DeviateNoise(d))
 
 
 @timer
@@ -1649,217 +1262,9 @@ def test_distLookupTable():
             test_array, np.array(dLookupTableResult), precisionF,
             err_msg='Wrong DistDeviate random number sequence from generate.')
 
-    # Test filling an image
-    d.seed(testseed)
-    testimage = galsim.ImageD(np.zeros((3, 1)))
-    testimage.addNoise(galsim.DeviateNoise(d))
-    np.testing.assert_array_almost_equal(
-            testimage.array.flatten(), np.array(dLookupTableResult), precision,
-            err_msg='Wrong DistDeviate random number sequence generated when applied to image.')
-
     # Check picklability
     do_pickle(d, lambda x: (x(), x(), x(), x()))
-    do_pickle(galsim.DeviateNoise(d), drawNoise)
     do_pickle(d)
-    do_pickle(galsim.DeviateNoise(d))
-
-
-@timer
-def test_ccdnoise():
-    """Test CCD Noise generator
-    """
-    for i in range(4):
-        prec = eval("precision"+typestrings[i])
-        cResult = eval("cResult"+typestrings[i])
-
-        rng = galsim.BaseDeviate(testseed)
-        ccdnoise = galsim.CCDNoise(rng, gain=cGain, read_noise=cReadNoise)
-        testImage = galsim.Image((np.zeros((2, 2))+sky).astype(types[i]))
-        ccdnoise.applyTo(testImage)
-        np.testing.assert_array_almost_equal(
-                testImage.array, cResult, prec,
-                err_msg="Wrong CCD noise random sequence generated for Image"+typestrings[i]+".")
-
-        # Check that reseeding the rng reseeds the internal deviate in CCDNoise
-        rng.seed(testseed)
-        testImage.fill(sky)
-        ccdnoise.applyTo(testImage)
-        np.testing.assert_array_almost_equal(
-                testImage.array, cResult, prec,
-                err_msg="Wrong CCD noise random sequence generated for Image"+typestrings[i]+
-                " after seed")
-
-        # Check using addNoise
-        rng.seed(testseed)
-        testImage.fill(sky)
-        testImage.addNoise(ccdnoise)
-        np.testing.assert_array_almost_equal(
-                testImage.array, cResult, prec,
-                err_msg="Wrong CCD noise random sequence generated for Image"+typestrings[i]+
-                " using addNoise")
-
-        # Test filling an image with Fortran ordering
-        rng.seed(testseed)
-        testImageF = galsim.Image(np.zeros((2, 2)).T, dtype=types[i])
-        testImageF.fill(sky)
-        testImageF.addNoise(ccdnoise)
-        np.testing.assert_array_almost_equal(
-                testImageF.array, cResult, prec,
-                err_msg="Wrong CCD noise generated for Fortran-ordered Image"+typestrings[i])
-
-        # Now include sky_level in ccdnoise
-        rng.seed(testseed)
-        ccdnoise = galsim.CCDNoise(rng, sky_level=sky, gain=cGain, read_noise=cReadNoise)
-        testImage.fill(0)
-        ccdnoise.applyTo(testImage)
-        np.testing.assert_array_almost_equal(
-                testImage.array, cResult-sky, prec,
-                err_msg="Wrong CCD noise random sequence generated for Image"+typestrings[i]+
-                " with sky_level included in noise")
-
-        rng.seed(testseed)
-        testImage.fill(0)
-        testImage.addNoise(ccdnoise)
-        np.testing.assert_array_almost_equal(
-                testImage.array, cResult-sky, prec,
-                err_msg="Wrong CCD noise random sequence generated for Image"+typestrings[i]+
-                " using addNoise with sky_level included in noise")
-
-    # Check CCDNoise variance:
-    var1 = sky/cGain + (cReadNoise/cGain)**2
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), var1, precision,
-            err_msg="CCDNoise getVariance returns wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getSkyLevel(), sky, precision,
-            err_msg="CCDNoise getSkyLevel returns wrong value")
-    np.testing.assert_almost_equal(
-            ccdnoise.getGain(), cGain, precision,
-            err_msg="CCDNoise getGain returns wrong value")
-    np.testing.assert_almost_equal(
-            ccdnoise.getReadNoise(), cReadNoise, precision,
-            err_msg="CCDNoise getReadNoise returns wrong value")
-
-    # Check that the noise model really does produce this variance.
-    # NB. If default float32 is used here, older versions of numpy will compute the variance
-    # in single precision, and with 2048^2 values, the final answer comes out significantly
-    # wrong (19.33 instead of 19.42, which gets compared to the nominal value of 19.44).
-    big_im = galsim.Image(2048,2048,dtype=float)
-    big_im.addNoise(ccdnoise)
-    var = np.var(big_im.array)
-    print('variance = ',var)
-    print('getVar = ',ccdnoise.getVariance())
-    np.testing.assert_almost_equal(
-            var, ccdnoise.getVariance(), 1,
-            err_msg='Realized variance for CCDNoise did not match getVariance()')
-
-    # Check that CCDNoise adds to the image, not overwrites the image.
-    gal = galsim.Exponential(half_light_radius=2.3, flux=0.3)
-    # Note: again, flux/size^2 needs to be << sky_level or it will mess up the statistics.
-    gal.drawImage(image=big_im)
-    big_im.addNoise(ccdnoise)
-    gal.withFlux(-0.3).drawImage(image=big_im, add_to_image=True)
-    var = np.var(big_im.array)
-    np.testing.assert_almost_equal(
-            var, ccdnoise.getVariance(), 1,
-            err_msg='CCDNoise wrong when already an object drawn on the image')
-
-    # Check withVariance
-    ccdnoise = galsim.CCDNoise(rng, sky_level=sky, gain=cGain, read_noise=cReadNoise)
-    ccdnoise = ccdnoise.withVariance(9.)
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 9., precision,
-            err_msg="CCDNoise withVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getSkyLevel(), (9./var1)*sky, precision,
-            err_msg="CCDNoise withVariance results in wrong SkyLevel")
-    np.testing.assert_almost_equal(
-            ccdnoise.getGain(), cGain, precision,
-            err_msg="CCDNoise withVariance results in wrong Gain")
-    np.testing.assert_almost_equal(
-            ccdnoise.getReadNoise(), np.sqrt(9./var1) * cReadNoise, precision,
-            err_msg="CCDNoise withVariance results in wrong ReadNoise")
-
-    # Check withScaledVariance
-    ccdnoise = ccdnoise.withScaledVariance(4.)
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 36., precision,
-            err_msg="CCDNoise withVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getSkyLevel(), (36./var1)*sky, precision,
-            err_msg="CCDNoise withVariance results in wrong SkyLevel")
-    np.testing.assert_almost_equal(
-            ccdnoise.getGain(), cGain, precision,
-            err_msg="CCDNoise withVariance results in wrong Gain")
-    np.testing.assert_almost_equal(
-            ccdnoise.getReadNoise(), np.sqrt(36./var1) * cReadNoise, precision,
-            err_msg="CCDNoise withVariance results in wrong ReadNoise")
-
-    # Check arithmetic
-    ccdnoise = ccdnoise.withVariance(0.5)
-    ccdnoise2 = ccdnoise * 3
-    np.testing.assert_almost_equal(
-            ccdnoise2.getVariance(), 1.5, precision,
-            err_msg="CCDNoise ccdnoise*3 results in wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 0.5, precision,
-            err_msg="CCDNoise ccdnoise*3 results in wrong variance for original ccdnoise")
-    ccdnoise2 = 5 * ccdnoise
-    np.testing.assert_almost_equal(
-            ccdnoise2.getVariance(), 2.5, precision,
-            err_msg="CCDNoise 5*ccdnoise results in wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 0.5, precision,
-            err_msg="CCDNoise 5*ccdnoise results in wrong variance for original ccdnoise")
-    ccdnoise2 = ccdnoise/2
-    np.testing.assert_almost_equal(
-            ccdnoise2.getVariance(), 0.25, precision,
-            err_msg="CCDNoise ccdnoise/2 results in wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 0.5, precision,
-            err_msg="CCDNoise 5*ccdnoise results in wrong variance for original ccdnoise")
-    ccdnoise *= 3
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 1.5, precision,
-            err_msg="CCDNoise ccdnoise*=3 results in wrong variance")
-    ccdnoise /= 2
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 0.75, precision,
-            err_msg="CCDNoise ccdnoise/=2 results in wrong variance")
-
-    # Check starting with CCDNoise()
-    ccdnoise = galsim.CCDNoise()
-    ccdnoise = ccdnoise.withVariance(9.)
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 9., precision,
-            err_msg="CCDNoise().withVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getSkyLevel(), 9., precision,
-            err_msg="CCDNoise().withVariance results in wrong SkyLevel")
-    np.testing.assert_almost_equal(
-            ccdnoise.getGain(), 1., precision,
-            err_msg="CCDNoise().withVariance results in wrong Gain")
-    np.testing.assert_almost_equal(
-            ccdnoise.getReadNoise(), 0., precision,
-            err_msg="CCDNoise().withVariance results in wrong ReadNoise")
-    ccdnoise = ccdnoise.withScaledVariance(4.)
-    np.testing.assert_almost_equal(
-            ccdnoise.getVariance(), 36., precision,
-            err_msg="CCDNoise().withScaledVariance results in wrong variance")
-    np.testing.assert_almost_equal(
-            ccdnoise.getSkyLevel(), 36., precision,
-            err_msg="CCDNoise().withScaledVariance results in wrong SkyLevel")
-    np.testing.assert_almost_equal(
-            ccdnoise.getGain(), 1., precision,
-            err_msg="CCDNoise().withScaledVariance results in wrong Gain")
-    np.testing.assert_almost_equal(
-            ccdnoise.getReadNoise(), 0., precision,
-            err_msg="CCDNoise().withScaledVariance results in wrong ReadNoise")
-
-    # Check picklability
-    do_pickle(ccdnoise, lambda x: (x.rng.serialize(), x.sky_level, x.gain, x.read_noise))
-    do_pickle(ccdnoise, drawNoise)
-    do_pickle(ccdnoise)
 
 
 @timer
@@ -1933,46 +1338,6 @@ def test_multiprocess():
 
 
 @timer
-def test_addnoisesnr():
-    """Test that addNoiseSNR is behaving sensibly.
-    """
-    # Rather than reproducing the S/N calculation in addNoiseSNR(), we'll just check for
-    # self-consistency of the behavior with / without flux preservation.
-    # Begin by making some object that we draw into an Image.
-    gal_sigma = 3.7
-    pix_scale = 0.6
-    test_snr = 73.
-    gauss = galsim.Gaussian(sigma=gal_sigma)
-    im = gauss.drawImage(scale=pix_scale, dtype=np.float64)
-
-    # Now make the noise object to use.
-    # Use a default-constructed rng (i.e. rng=None) since we had initially had trouble
-    # with that.  And use the duplicate feature to get a second copy of this rng.
-    gn = galsim.GaussianNoise()
-    rng2 = gn.getRNG().duplicate()
-
-    # Try addNoiseSNR with preserve_flux=True, so the RNG needs a different variance.
-    # Check what variance was added for this SNR, and that the RNG still has its original variance
-    # after this call.
-    var_out = im.addNoiseSNR(gn, test_snr, preserve_flux=True)
-    assert gn.getVariance()==1.0
-    max_val = im.array.max()
-
-    # Now apply addNoiseSNR to another (clean) image with preserve_flux=False, so we use the noise
-    # variance in the original RNG, i.e., 1.  Check that the returned variance is 1, and that the
-    # value of the maximum pixel (presumably the peak of the galaxy light profile) is scaled as we
-    # expect for this SNR.
-    im2 = gauss.drawImage(scale=pix_scale, dtype=np.float64)
-    gn2 = galsim.GaussianNoise(rng=rng2)
-    var_out2 = im2.addNoiseSNR(gn2, test_snr, preserve_flux=False)
-    assert var_out2==1.0
-    expect_max_val2 = max_val*np.sqrt(var_out2/var_out)
-    np.testing.assert_almost_equal(
-            im2.array.max(), expect_max_val2, decimal=8,
-            err_msg='addNoiseSNR with preserve_flux = True and False give inconsistent results')
-
-
-@timer
 def test_permute():
     """Simple tests of the permute() function."""
     # Make a fake list, and another list consisting of indices.
@@ -2029,8 +1394,6 @@ if __name__ == "__main__":
     test_chi2()
     test_distfunction()
     test_distLookupTable()
-    test_ccdnoise()
     test_multiprocess()
-    test_addnoisesnr()
     test_permute()
     test_ne()
